프로젝트 구조 변경에 의한 Onepage Test JS, Java Code 수정
Heroku Plugin 제거
불필요한 파일 삭제 및 .gitignore 적용
