+++ .gitignore
@@ -3,3 +3,7 @@
 /.settings
 /.project
 /.classpath
+.DS_Store
+Procfile
+app.json
+system.properties
