--- Procfile
@@ -1,1 +1,0 @@
-web: sh target/bin/kollus-sample-jsp
