--- app.json
@@ -1,11 +1,0 @@
-{
-  "name": "kollus-sample-jsp",
-  "description": "kollus-sample-jsp",
-  "repository": "https://git.heroku.com/kollus-sample-jsp",
-  "logo": "https://upload.wikimedia.org/wikipedia/commons/7/7b/Tomcat-logo.svg",
-  "keywords": [
-    "java",
-    "tomcat",
-    "kollus"
-  ]
-}
