+++ pom.xml
@@ -91,15 +91,6 @@
 		<finalName>kollus-sample-jsp</finalName>
 		<plugins>
 			<plugin>
-				<groupId>com.heroku.sdk</groupId>
-				<artifactId>heroku-maven-plugin</artifactId>
-				<version>1.1.3</version>
-				<configuration>
-					<appName>kollus-sample-servlet</appName>
-					<jdkVersion>1.8</jdkVersion>
-				</configuration>
-			</plugin>
-			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-compiler-plugin</artifactId>
 				<version>3.1</version>
