+++ src/main/java/com/kollus/kr/kollus_sample_java/controller/MediaServlet.java
@@ -19,6 +19,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.google.gson.Gson;
 import com.google.gson.JsonElement;
 import com.jayway.jsonpath.JsonPath;
+import com.kollus.kr.kollus_sample_java.config.UrlConf;
 import com.kollus.kr.kollus_sample_java.config.UserConf;
 import com.kollus.kr.kollus_sample_java.util.HttpUtil;
 import com.kollus.kr.kollus_sample_java.util.JwtUtil;
@@ -69,34 +70,24 @@ public class MediaServlet extends HttpServlet {
 		if (subPaths.length < 2) {
 			/* TODO /media/ */
 		} else if (subPaths.length == 2) {
-			if ("category".equals(subPaths[1]) || "channel".equals(subPaths[1])) {
-				_url = String.format("http://api.kr.kollus.com/0/media/%s/media_content?access_token=%s&per_page=%s",
-						subPaths[1], accessToken, page);
+			if ("category".equals(subPaths[1])){
+				_url = String.format(UrlConf.KOLLUS_LIBRARY_ALL_CONTENTS_URL, accessToken, page);
+			}
+			else if( "channel".equals(subPaths[1])) {
+				_url = String.format(UrlConf.KOLLUS_CHANNEL_ALL_CONTENTS_URL, accessToken, page);
 			} else {
 			}
 
-		} else if (subPaths.length == 3) {
-			if ("category".equals(subPaths[1])) {
-				_url = String.format(
-						"http://api.kr.kollus.com/0/media/category/media_content?access_token=%s&category_key=%s&per_page=%s",
-						accessToken, subPaths[2], page);
-			} else if ("channel".equals(subPaths[1])) {
-				_url = String.format(
-						"http://api.kr.kollus.com/0/media/channel/media_content?access_token=%s&channel_key=%s&per_page=%s",
-						accessToken, subPaths[2], page);
+		}
+		else if (subPaths.length == 3) {
+			if ("category".equals(subPaths[1])){
+				_url = String.format(UrlConf.KOLLUS_LIBRARY_CONTENTS_URL, accessToken, subPaths[2],page);
+			}
+			else if( "channel".equals(subPaths[1])) {
+				_url = String.format(UrlConf.KOLLUS_CHANNEL_CONTENTS_URL, accessToken, subPaths[2],page);
 			} else {
 			}
-		} else if (subPaths.length == 4) {
-			if ("category".equals(subPaths[1])) {
-				_url = String.format(
-						"http://api.kr.kollus.com/0/media/category/media_content?access_token=%s&category_key=%s&per_page=%s",
-						accessToken, subPaths[2], page);
-			} else if ("channel".equals(subPaths[1])) {
-				_url = String.format(
-						"http://api.kr.kollus.com/0/media/category/media_content?access_token=%s&channel_key=%s&per_page=%s",
-						accessToken, subPaths[2], page);
-			} else {
-			}
+
 		}
 		String result = null;
 		List<HashMap<String, Object>> resultMap = null;
