+++ src/main/webapp/files/user.properties
@@ -1,5 +1,5 @@
-#Fri Jul 07 10:04:12 KST 2017
-secretKey=
-accountKey=
-accessToken=
-userKey=
+#Fri Jul 07 11:21:05 KST 2017
+secretKey=hdyang2
+accountKey=hdyang2
+accessToken=7ge80tfvz51x2606
+userKey=3463b89d188a9bfa727088ffae2af078754c7f86a0d56b65a5cf59df98250eb5
