--- system.properties
@@ -1,2 +1,0 @@
-java.runtime.version=1.8.0_121
-maven.version=3.3.9
